+++ src/main/java/com/epam/reportportal/cucumber/AbstractReporter.java
@@ -436,8 +436,7 @@ public abstract class AbstractReporter implements Formatter {
 		String errorMessage = result.getErrorMessage();
 		if (errorMessage != null) {
 			sendLog(errorMessage, level);
-		}
-		if (result.getError() != null) {
+		} else if (result.getError() != null) {
 			sendLog(getStackTraceAsString(result.getError()), level);
 		}
 	}