+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Added
 - Docstring parameter handling
+### Changed
+- Many static methods from Util class were moved to AbstractReporter class and made protected to ease extension 
 
 ## [5.0.0-RC-1]
 ### Added