+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Added
 - Missed javadocs and annotations
 - A protected method `buildStartFeatureRequest` to ease Feature creation customization
+### Changed
+- Client version updated on [5.0.15](https://github.com/reportportal/client-java/releases/tag/5.0.15)
 
 ## [5.0.0]
 ### Added