+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - A protected method `buildStartFeatureRequest` to ease Feature creation customization
 ### Changed
 - Client version updated on [5.0.15](https://github.com/reportportal/client-java/releases/tag/5.0.15)
+### Fixed
+- 'CHILD_START_TIME_EARLIER_THAN_PARENT' Exception in some cases
 
 ## [5.0.0]
 ### Added