+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Client version updated on [5.0.15](https://github.com/reportportal/client-java/releases/tag/5.0.15)
 ### Fixed
 - 'CHILD_START_TIME_EARLIER_THAN_PARENT' Exception in some cases
+- Double error message reporting
 
 ## [5.0.0]
 ### Added