+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Missed javadocs and annotations
+- A protected method `buildStartFeatureRequest` to ease Feature creation customization
+### Changed
+- Client version updated on [5.0.15](https://github.com/reportportal/client-java/releases/tag/5.0.15)
+### Fixed
+- 'CHILD_START_TIME_EARLIER_THAN_PARENT' Exception in some cases
+- Double error message reporting
 
 ## [5.0.0]
 ### Added